Fix migration drop column issue

+++ database/migrations/2019_06_03_110340_add_sid_to_emails_table.php
@@ -25,8 +25,10 @@ class AddSidToEmailsTable extends Migration
      */
     public function down()
     {
-        Schema::table('emails', function (Blueprint $table) {
-            $table->dropColumn('sid');
-        });
+        if (Schema::hasColumn('emails', 'sid')) {
+            Schema::table('emails', function (Blueprint $table) {
+                $table->dropColumn('sid');
+            });
+        }
     }
 }
